+++ modules/eq_pentaho_translate/__openerp__.py
@@ -36,10 +36,10 @@
     # Check https://github.com/odoo/odoo/blob/master/openerp/addons/base/module/module_data.xml
     # for the full list
     'category': 'Uncategorized',
-    'version': '1.0.0',
+    'version': '1.0.1',
     'license': 'AGPL-3',
     # any module necessary for this one to work correctly
-    'depends': ['base','product'],
+    'depends': ['base','product', 'sale'],
 
     # always loaded
     'data': [